requirements.txt
google-genai>=1.0.0
chromadb>=1.0.0
rank-bm25>=0.2.2
chainlit>=1.1.0
python-dotenv>=1.0.0
pymupdf>=1.24.0
